Fix h1 reports 3653196 + 3653316

+++ www/admin/zone-include.php
@@ -39,11 +39,11 @@ phpAds_SessionDataStore();
 
 // Get input parameters
 $pref = &$GLOBALS['_MAX']['PREF'];
-$publisherId = MAX_getValue('affiliateid');
-$zoneId = MAX_getValue('zoneid');
-$advertiserId = MAX_getValue('clientid');
-$placementId = MAX_getValue('campaignid');
-$adId = MAX_getValue('bannerid');
+$publisherId = (int) MAX_getValue('affiliateid');
+$zoneId = (int) MAX_getValue('zoneid');
+$advertiserId = (int) MAX_getValue('clientid');
+$placementId = (int) MAX_getValue('campaignid');
+$adId = (int) MAX_getValue('bannerid');
 $action = MAX_getValue('action');
 $aCurrent = MAX_getValue('includebanner');
 $hideInactive = MAX_getStoredValue('hideinactive', ($pref['ui_hide_inactive'] == true), null, true);
